requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
cffi==1.15.0
cryptography==36.0.2
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
pytz==2022.1
sqlparse==0.4.2
